Fix for issue #24 - double import - relocate files from pyftdi.pyftdi to pyftdi - adjust imports from pyftdi.pyftdi.x to pyftdi.x - fix setup.py to remove pyftdi.pyftdi
