+++ pyftdi/misc.py
