+++ pyftdi/tests/bits.py
